xyz/requirements.txt
openai
python-dotenv
pydantic
fastapi
pandas